ai-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dateutil
parsedatetime

